@@ -1,6 +1,6 @@
 {
   "name": "not-store",
-  "version": "3.1.5",
+  "version": "3.1.6",
   "description": "file storage interface",
   "main": "index.cjs",
   "repository": "https://github.com/interrupter/not-store",
@@ -18,7 +18,7 @@
     "test:server:off": "wget -qO- http://localhost:7357/terminate",
     "test": "npm run test:server & npm run test:node && npm run test:run",
     "testOnly": "npm run test:server & sleep 4 && npm run test:node && npm run test:run && npm run test:server:off",
-    "lint": "/usr/local/bin/eslint --ext svelte3 ./src/**/**/**/**",
+    "lint": "./node_modules/.bin/eslint --ext svelte3 ./src/**/**/**/**",
     "pretest": "./node_modules/.bin/eslint ./src/**/**",
     "docs": "./node_modules/.bin/jsdoc -c jsdoc.json",
     "cover": "npm run test:server & sleep 4 && npm run cover:node && npm run cover:run && npm run cover:report",
@@ -41,15 +41,15 @@
   "license": "ISC",
   "dependencies": {
     "@shopify/draggable": "^1.1.3",
-    "aws-sdk": "^2.1691.0",
+    "aws-sdk": "^2.1692.0",
     "download": "^8.0.0",
     "easy-yandex-s3": "^2.0.0",
     "is-stream": "2.*",
-    "jsdoc": "^4.0.3",
+    "jsdoc": "^4.0.4",
     "mkdirp": "*",
     "not-error": "^0.3.0",
     "not-path": "^1.0.6",
-    "not-validation": "^0.0.9",
+    "not-validation": "^0.0.11",
     "sharp": "^0.33.5",
     "streamifier": "^0.1.1",
     "valid-url": "^1.0.9"
@@ -57,7 +57,7 @@
   "devDependencies": {
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-syntax-class-properties": "^7.12.13",
-    "@cypress/code-coverage": "^3.13.2",
+    "@cypress/code-coverage": "^3.13.7",
     "babel-eslint": "^10.1.0",
     "babel-plugin-istanbul": "^7.0.0",
     "chai": "*",
@@ -65,34 +65,34 @@
     "connect": "^3.7.0",
     "connect-mongo": "^5.1.0",
     "cors": "^2.8.5",
-    "cypress": "^13.14.2",
+    "cypress": "^13.16.0",
     "cypress-file-upload": "^5.0.8",
     "eslint": "*",
-    "eslint-plugin-perfectionist": "^3.7.0",
-    "express": "^4.21.0",
+    "eslint-plugin-perfectionist": "^4.0.3",
+    "express": "^4.21.1",
     "express-fileupload": "^1.5.1",
-    "express-session": "^1.18.0",
-    "helmet": "^7.1.0",
-    "husky": "^9.1.6",
+    "express-session": "^1.18.1",
+    "helmet": "^8.0.0",
+    "husky": "^9.1.7",
     "ink-docstrap": "^1.3.2",
     "ioredis": "^5.4.1",
     "method-override": "^3.0.0",
     "mocha": "*",
     "mockgoose": "^8.0.4",
-    "mongodb-memory-server": "^10.0.1",
-    "mongoose": "^8.6.3",
-    "not-bulma": "^1.2.68",
+    "mongodb-memory-server": "^10.1.2",
+    "mongoose": "^8.8.2",
+    "not-bulma": "^2.0.2",
     "not-config": "^0.1.8",
     "not-log": "^0.0.21",
     "not-meta": "^0.0.28",
-    "not-node": "^6.5.27",
-    "not-options": "^0.2.7",
-    "not-user": "^0.2.162",
+    "not-node": "^6.5.37",
+    "not-options": "^0.3.0",
+    "not-user": "^1.0.2",
     "nyc": "^17.1.0",
-    "postcss": "^8.4.47",
+    "postcss": "^8.4.49",
     "redis": "^4.7.0",
     "request": "^2.88.2",
-    "retire": "^5.2.3",
+    "retire": "^5.2.5",
     "@rollup/plugin-babel": "*",
     "@rollup/plugin-commonjs": "*",
     "rollup-plugin-filesize": "^10.0.0",
@@ -100,11 +100,11 @@
     "@rollup/plugin-node-resolve": "*",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-svelte": "^7.2.2",
-    "superagent": "^10.1.0",
-    "svelte": ">=4.2.19",
-    "svelte-dnd-action": "^0.9.50",
+    "superagent": "^10.1.1",
+    "svelte": ">=5.2.7",
+    "svelte-dnd-action": "^0.9.52",
     "taffydb": "^2.7.3",
-    "uuid": "^10.0.0"
+    "uuid": "^11.0.3"
   },
   "nyc": {
     "include": [
