@@ -1,19 +1,18 @@
 // @ts-check
 const notStoreProcessor = require("../../../proto/processor.cjs");
-const getApp = require('not-node/src/getApp');
+const getApp = require("not-node/src/getApp");
 const DEFAULT_OPTIONS = require("./image.parent.preview.options.cjs");
 
 /**
- * После загрузки изображения, если это вторичное изображение, 
+ * После загрузки изображения, если это вторичное изображение,
  * и variant === установленному в настройках для процессора,
  * то этот файл считается "превью" для первичного файла и для него
  *  устанавливается ссылка на превью из cloud.Location
- * 
+ *
  * @class notStoreProcessorImageParentPreview
  * @extends {notStoreProcessor}
  */
 class notStoreProcessorImageParentPreview extends notStoreProcessor {
-    
     static getOptions() {
         return JSON.parse(JSON.stringify(DEFAULT_OPTIONS));
     }
@@ -22,8 +21,8 @@ class notStoreProcessorImageParentPreview extends notStoreProcessor {
         return {
             id: "image.parent.preview",
             title: "Установка превью для основного файла",
-            optionsDefault: this.getOptions(), 
-            optionsUI: "UIStoreProcessorOptionsImageUploadParentPreview",  
+            optionsDefault: this.getOptions(),
+            optionsUI: "UIStoreProcessorOptionsImageUploadParentPreview",
         };
     }
 
@@ -33,13 +32,14 @@ class notStoreProcessorImageParentPreview extends notStoreProcessor {
      * @static
      * @param {object} file
      * @param {object} options
-     * @param {import('../../../drivers/timeweb/timeweb.driver.cjs')} driver
      * @returns {Promise<undefined>}
      * @memberof notStoreProcessorImageUpload
      */
-    static async run(file, options, driver) {
-        if(file.parent && (options.all || file.variant === options.variant)){
-            await getApp().getModel('File').setVariantURL(file.parent, file.variant, file.cloud.Location);
+    static async run(file, options) {
+        if (file.parent && (options.all || file.variant === options.variant)) {
+            await getApp()
+                .getModel("File")
+                .setVariantURL(file.parent, file.variant, file.cloud.Location);
         }
     }
 }
